refactor(cli): split cmd/file-search-on/main.go into per-subcommand files (closes #155)
main.go had accumulated 1209 LOC across 14 subcommands. Every other
multi-format family in the codebase has been split when it grew this
dense (disk_*.go #127, install_*.go #131, bytecode_*.go #140,
body_*.go #141), and hashset_cmd.go (#150) was already the
per-subcommand-file precedent inside cmd/. This extends that to the
remaining 13 subcommands.

Pure refactor — no behaviour change. Mechanical move of each
subcommand's struct + Run method into its own file; main.go becomes
the kong root + main() + truly-shared helpers.

File layout
===========

  main.go              137 LOC  CLI root, main(), openIndex,
                                exitCodeError, isCancellation
  attrs_cmd.go          65 LOC  AttrsCmd
  lines_cmd.go          50 LOC  LinesCmd
  stats_cmd.go          93 LOC  StatsCmd
  duplicates_cmd.go     91 LOC  DuplicatesCmd
  near_duplicates_cmd.go 93 LOC NearDuplicatesCmd
  archive_cmd.go       118 LOC  ArchiveContentsCmd + ArchiveReadCmd
  find_matches_cmd.go  110 LOC  FindMatchesCmd
  project_cmd.go       143 LOC  DetectProjectCmd + WhichProjectCmd +
                                FindProjectsCmd + ConfigPathsCmd
  mcp_cmd.go            46 LOC  MCPCmd
  search_cmd.go        382 LOC  SearchCmd + streamSearch +
                                bufferedSearch + printHelp +
                                printAttrs + printFuncs
  output.go            715 LOC  (unchanged — printer dispatch)
  hashset_cmd.go        99 LOC  (unchanged — already extracted in #150)

main.go shrinks from 1209 → 137 LOC (88% reduction).

Verification
============

  go build ./...        clean
  go vet ./...          clean
  go fix -diff ./...    empty
  go test -race ./...   all packages green
  golangci-lint run     0 issues

End-to-end spot checks (file-search-on --help / file-search-on mcp
--help / file-search-on hash-set --help): identical output to
pre-refactor.
