+++ cmd/file-search-on/attrs_cmd.go
@@ -1,0 +1,65 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/richardwooding/file-search-on/internal/celexpr"
+	contentpkg "github.com/richardwooding/file-search-on/internal/content"
+	"github.com/richardwooding/file-search-on/internal/search"
+)
+
+type AttrsCmd struct {
+	Path   string `arg:"" help:"File to inspect."`
+	Output string `short:"o" name:"output" enum:"default,verbose,json" default:"verbose" help:"Output format: default | verbose | json."`
+	Format string `name:"format" help:"Custom Go text/template applied to the record (e.g. '{{.Path}}\\t{{.Title}}'). When set, takes precedence over -o."`
+}
+
+func (a *AttrsCmd) Run(ctx context.Context) error {
+	abs, err := filepath.Abs(a.Path)
+	if err != nil {
+		return fmt.Errorf("resolve path: %w", err)
+	}
+	info, err := os.Stat(abs)
+	if err != nil {
+		return fmt.Errorf("stat %s: %w", abs, err)
+	}
+	if info.IsDir() {
+		return fmt.Errorf("%s is a directory; use the search subcommand to walk a tree", abs)
+	}
+
+	dir := filepath.Dir(abs)
+	base := filepath.Base(abs)
+
+	attrs, err := celexpr.BuildAttributes(ctx, os.DirFS(dir), base, abs, contentpkg.DefaultRegistry())
+	if err != nil {
+		return fmt.Errorf("read attributes: %w", err)
+	}
+
+	result := search.Result{
+		Path:        abs,
+		ContentType: attrs.ContentType,
+		Size:        attrs.Size,
+		Attrs:       attrs,
+	}
+	results := []search.Result{result}
+
+	if a.Format != "" {
+		tmpl, err := parseFormatTemplate(a.Format)
+		if err != nil {
+			return fmt.Errorf("parse --format template: %w", err)
+		}
+		return printTemplate(os.Stdout, results, tmpl)
+	}
+	switch a.Output {
+	case "json":
+		return printJSON(os.Stdout, results)
+	case "default":
+		printDefault(os.Stdout, results)
+	default: // "" or "verbose"
+		printVerbose(os.Stdout, results)
+	}
+	return nil
+}
